fix code

+++ src/coding/algorithm.cpp
@@ -9,9 +9,9 @@
 #include "iostream"
 namespace cyy::algorithm {
   std::unordered_map<symbol_type, std::string>
-  huffman_code(const std::unordered_map<symbol_type, float> &symbol_frequency) {
+  huffman_code(const std::unordered_map<symbol_type, double> &symbol_frequency) {
     assert(!symbol_frequency.empty());
-    heap<symbol_type, float> h;
+    heap<symbol_type, double> h;
     for (auto [symbol, frequency] : symbol_frequency) {
       h.insert(symbol, frequency);
     }
